btf: add function type support
This patch adds BTF_KIND_FUNC and BTF_KIND_FUNC_PROTO
support to the type section. The BTF_KIND_FUNC is the type
for dwarf subprogram tag which represents a defined subprogram,
and the BTF_KIND_FUNC_PROTO is the type for dwarf subroutine_type
which represents the pointee of a function pointer.

The func return type is in t->type (where t is a "struct btf_type"
object).  The func args are an array of u32s immediately
following object "t".

Signed-off-by: Okash Khawaja <osk@fb.com>
Signed-off-by: Yonghong Song <yhs@fb.com>